# Segment 2D images without patches