--- a/ivadomed/main.py
+++ b/ivadomed/main.py
@@ -69,6 +69,11 @@ def get_parser():
     optional_args.add_argument('--resume-training', dest="resume_training", required=False, action='store_true',
                                help='Load a saved model ("checkpoint.pth.tar" in the output directory specified either with flag "--path-output" or via the config file "output_path" argument)  '
                                     'for resume training. This training state is saved everytime a new best model is saved in the output directory specified with flag "--path-output"')
+    optional_args.add_argument('--patch', action='store_true', required=False, help='Uses 2D patches while segmenting '
+                               '(command "--segment").')
+    optional_args.add_argument('--overlap-2d', dest="overlap_2d", required=False, type=int, nargs="+", help='Custom '
+                               'overlap for 2D patches while segmenting (command "--segment"). Default model overlap '
+                               'is used otherwise.')
     optional_args.add_argument('-h', '--help', action='help', default=argparse.SUPPRESS,
                                help='Shows function documentation.')
 
@@ -223,7 +228,7 @@ def update_film_model_params(context, ds_test, model_params, path_output):
     return ds_test, model_params
 
 
-def run_segment_command(context, model_params):
+def run_segment_command(context, model_params, patch, overlap_2d):
     # BIDSDataframe of all image files
     # Indexing of derivatives is False for command segment
     bids_df = BidsDataframe(
@@ -289,6 +294,11 @@ def run_segment_command(context, model_params):
             options[OptionKW.PIXEL_SIZE_UNITS] = \
                 bids_df.df.loc[bids_df.df['filename'] == subject][MetadataKW.PIXEL_SIZE_UNITS].values[0]
 
+        # Add 'patch' and 'overlap-2d' argument to options
+        options['patch'] = patch
+        if overlap_2d:
+            options['overlap_2D'] = overlap_2d
+
         if fname_img:
             pred_list, target_list = imed_inference.segment_volume(str(path_model),
                                                                    fname_images=fname_img,
@@ -314,7 +324,7 @@ def run_segment_command(context, model_params):
                                            suffix="_pred.png")
 
 
-def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
+def run_command(context, n_gif=0, thr_increment=None, resume_training=False, patch=False, overlap_2d=None):
     """Run main command.
 
     This function is central in the ivadomed project as training / testing / evaluation commands
@@ -329,8 +339,13 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
         thr_increment (float): A threshold analysis is performed at the end of the training using the trained model and
             the training + validation sub-dataset to find the optimal binarization threshold. The specified value
             indicates the increment between 0 and 1 used during the ROC analysis (e.g. 0.1).
-        resume_training (bool): Load a saved model ("checkpoint.pth.tar" in the output directory specified with flag "--path-output" or via the config file "output_path" '            This training state is saved everytime a new best model is saved in the log
-            argument) for resume training directory.
+        resume_training (bool): Load a saved model ("checkpoint.pth.tar" in the output directory specified with flag
+            "--path-output" or via the config file "output_path" ' This training state is saved everytime a new best
+            model is saved in the log argument) for resume training directory.
+        patch (bool): Indicate if 2d patching is used while segmenting (command '--segment'), otherwise inference is
+            performed on the entire image at once. Default: False.
+        overlap_2d (list of int): Custom overlap for 2D patches while segmenting (command "--segment").
+            Default model overlap is used otherwise.
 
     Returns:
         float or pandas.DataFrame or None:
@@ -367,7 +382,7 @@ def run_command(context, n_gif=0, thr_increment=None, resume_training=False):
     model_params, loader_params = set_model_params(context, loader_params)
 
     if command == 'segment':
-        run_segment_command(context, model_params)
+        run_segment_command(context, model_params, patch, overlap_2d)
         return
 
     # BIDSDataframe of all image files
@@ -597,7 +612,9 @@ def run_main():
     run_command(context=context,
                 n_gif=args.gif if args.gif is not None else 0,
                 thr_increment=args.thr_increment if args.thr_increment else None,
-                resume_training=bool(args.resume_training))
+                resume_training=bool(args.resume_training),
+                patch=args.patch if args.patch is not None else False,
+                overlap_2d=args.overlap_2d if args.overlap_2d is not None else None)
 
 
 if __name__ == "__main__":