@hermancollin 2022-07-06
I can review this PR when it's ready. It would be nice to also offer this option in ADS. At a quick glance, I see you added an argument to `segment_volume` so I'm guessing it would be pretty easy to implement on the ADS side.