@mariehbourget 2022-07-06
> I see you added an argument to `segment_volume` so I'm guessing it would be pretty easy to implement on the ADS side.

Yes exactly, we would only have to pass the argument to segment_volume without changes to the config file.