@hermancollin 2022-07-12
@mariehbourget I will review this today but I don't understand how it works. My understanding was that u-nets have a fixed input size. Is the image resized to the input patch size the model was trained on? That doesn't make sense because your large segmented images in #1049 would never be as good. Or is the image subdivided in 512x512 patches (for example) without overlap and with padding when the size is smaller than the patch size?