@mariehbourget 2022-07-12
Hi @hermancollin, great questions!
My understanding may not be complete either on what is happening, but here is what I gathered from my tests and from the tests in #1049.

> Is the image resized to the input patch size the model was trained on?

No, the only resampling is the resampling to the common resolution (`Resample` param), wihtout any ties to the original patch size used.

> Or is the image subdivided in 512x512 patches (for example) without overlap and with padding when the size is smaller than the patch size?

No, there is no patches at all.
A model with patches and without overlap would have a `length_2D == stride_2D`.
In 2D in ivadomed, there is never any padding (I think), only for sub-volumes in 3D which I'm not sure why.

> My understanding was that u-nets have a fixed input size.

> I get that these networks should theoretically be able to segment any image size because the learned weights are in the convolutional kernels, but since I'm not as familiar with ivadomed codebase, I don't see the underlying logic.

From what I understand, the model doesn't care about the size of the input. As you said, the weights are in the convolutional kernels and they are shared by all the pixels of the same layer, regardless of how many pixels there are.

So if I'm not mistaken, the "fixed" input size is only really to be able to use a batch size > 1 i.e. combine several images in the same tensor of size (batchsize, channels, X, Y).
If you use a batch size of 1, I think you should be able to train a model with images of different sizes.
However, the "fixed" smaller patch sizes are also probably necessary to avoid memory issues with large images.

So we are able to segment different image sizes because:
- each image is loaded independently of the others i.e. batch_size = 1
- in the case of a 3D volume segmented by 2D slices, all slices have the same dimensions so the batch size is not important.

Let me know if that makes sense.