@hermancollin 2022-07-12
Yes that makes a lot of sense. Even though you reported the metrics were slightly worse when using `--no-patch`, I feel like they are more visually pleasing. The "patching artifacts" or patch-overlap problems we often see have a definite impact on the seg quality, especially for smaller axons. Without patches, I notice a bunch of small axons suddently appear out of thin air.

Everything LGTM. The option makes ivadomed load the image as a single slice. Also, it's faster than I expected. Using `segment_volume` directly (ONNX model):

with patches:
```
python script_patching.py  29,33s user 1,67s system 176% cpu 17,592 total
```
without patches:
```
python script_no_patching.py  17,69s user 0,93s system 169% cpu 10,984 total
```
Keep in mind the script with patching had some slight patch overlap.